comment LEPRIKON_DOMAIN by default
If you run docker-compose up with default config without your instance being exposed to the internet and your dns correctly setup then leprikon will fail to start because nginx expects certificates but there are none.

I have also changed the domain because if you read the logs and supply certificates manually the 404 page of leprikon.cz will be shown which is confusing, you may assume that something is wrong with your leprikon (it took me a while to realize this was redirect).

I have also added more information about this to README.md